ref(core): Extract cookie utilities from HTTP utilities
Move cookie parsing and filtering into a focused internal utility and update integrations to use it. Keep generic query and header filtering in HttpUtils.
