+++ sentry-ktor-client/src/main/java/io/sentry/ktorClient/SentryKtorClientUtils.kt
@@ -16,6 +16,7 @@ import io.sentry.TypeCheckHint
 import io.sentry.exception.ExceptionMechanismException
 import io.sentry.exception.SentryHttpClientException
 import io.sentry.protocol.Mechanism
+import io.sentry.util.CookieUtils
 import io.sentry.util.HttpUtils
 import io.sentry.util.UrlUtils
 
@@ -67,7 +68,7 @@ internal object SentryKtorClientUtils {
 
   private fun getRequestCookies(scopes: IScopes, cookies: String?): String? =
     if (scopes.options.dataCollectionResolver.isDataCollectionConfigured) {
-      HttpUtils.filterCookies(
+      CookieUtils.filterCookies(
         cookies,
         scopes.options.dataCollectionResolver.cookies,
         null,
@@ -80,7 +81,7 @@ internal object SentryKtorClientUtils {
 
   private fun getResponseCookies(scopes: IScopes, cookies: String?): String? =
     if (scopes.options.dataCollectionResolver.isDataCollectionConfigured) {
-      HttpUtils.filterSetCookie(cookies, scopes.options.dataCollectionResolver.cookies)
+      CookieUtils.filterSetCookie(cookies, scopes.options.dataCollectionResolver.cookies)
     } else if (scopes.options.isSendDefaultPii) {
       cookies
     } else {
