+++ sentry-opentelemetry/sentry-opentelemetry-core/src/main/java/io/sentry/opentelemetry/OpenTelemetryAttributesExtractor.java
@@ -9,6 +9,7 @@ import io.sentry.IScope;
 import io.sentry.SentryLevel;
 import io.sentry.SentryOptions;
 import io.sentry.protocol.Request;
+import io.sentry.util.CookieUtils;
 import io.sentry.util.HttpUtils;
 import io.sentry.util.StringUtils;
 import io.sentry.util.UrlUtils;
@@ -91,7 +92,7 @@ public final class OpenTelemetryAttributesExtractor {
                   headers.put(
                       headerName,
                       toString(
-                          HttpUtils.filterOutSecurityCookiesFromHeader(
+                          CookieUtils.filterOutSecurityCookiesFromHeader(
                               headerValues, headerName, null)));
                 } catch (Throwable t) {
                   options
