+++ sentry-spring/src/main/java/io/sentry/spring/webflux/SentryRequestResolver.java
@@ -3,6 +3,7 @@ package io.sentry.spring.webflux;
 import com.jakewharton.nopen.annotation.Open;
 import io.sentry.IScopes;
 import io.sentry.protocol.Request;
+import io.sentry.util.CookieUtils;
 import io.sentry.util.HttpUtils;
 import io.sentry.util.Objects;
 import io.sentry.util.UrlUtils;
@@ -37,18 +38,18 @@ public class SentryRequestResolver {
     urlDetails.applyToRequest(sentryRequest);
     sentryRequest.setHeaders(resolveHeadersMap(httpRequest.getHeaders()));
 
-    final @NotNull String headerName = HttpUtils.COOKIE_HEADER_NAME;
+    final @NotNull String headerName = CookieUtils.COOKIE_HEADER_NAME;
     if (scopes.getOptions().getDataCollectionResolver().isDataCollectionConfigured()) {
       sentryRequest.setCookies(
           toString(
-              HttpUtils.filterCookiesFromHeader(
+              CookieUtils.filterCookiesFromHeader(
                   httpRequest.getHeaders().get(headerName),
                   scopes.getOptions().getDataCollectionResolver().getCookies(),
                   Collections.emptyList())));
     } else if (scopes.getOptions().isSendDefaultPii()) {
       sentryRequest.setCookies(
           toString(
-              HttpUtils.filterOutSecurityCookiesFromHeader(
+              CookieUtils.filterOutSecurityCookiesFromHeader(
                   httpRequest.getHeaders().get(headerName), headerName, Collections.emptyList())));
     }
     return sentryRequest;
@@ -65,7 +66,7 @@ public class SentryRequestResolver {
         headersMap.put(
             headerName,
             toString(
-                HttpUtils.filterOutSecurityCookiesFromHeader(
+                CookieUtils.filterOutSecurityCookiesFromHeader(
                     entry.getValue(), headerName, Collections.emptyList())));
       }
     }
